Docker/backend/django_backend/requirements.txt
asgiref==3.8.1
click==8.1.7
Django==5.0.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dlib==19.24.6
face-recognition==1.3.0
face_recognition_models==0.3.0
filelock==3.15.4
fsspec==2024.6.1
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==2.0.1
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
pillow==10.4.0
psycopg2==2.9.9
PyJWT==2.8.0
scipy==1.14.0
setuptools==72.1.0
sqlparse==0.5.1
sympy==1.13.2
torch==2.4.0
torchvision==0.19.0
typing_extensions==4.12.2
wheel==0.44.0
